requirements.txt
Pillow==7.2.0
instabot==0.117.0
python-dotenv==0.9.1
requests==2.24.0